requirements.txt
     transformers
     lxml
     beautifulsoup4
     streamlit
     python-dotenv
     requests
     torch
     